requirements.txt
django==5.1.3
djangorestframework==3.15.2
environs==11.2.1
drf-spectacular==0.27.2

psycopg
psycopg[binary]